+++ core/src/com/mygdx/game/GameEntities/Vegetable.java
@@ -9,7 +9,7 @@ public class Vegetable extends Collectible {
         super(x, y, speed, texture, movement);
     }
 
-    public Vegetable(float x, float y, float speed, Texture texture, AIMovement movement, float gamePoints) {
+    public Vegetable(float x, float y, float speed, Texture texture, AIMovement movement, int gamePoints) {
         super(x, y, speed, texture, movement);
         this.setPoints(gamePoints);
     }
